% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamfuncs.R
\name{fun1}
\alias{fun1}
\title{Test function level 1}
\usage{
fun1(fun2.params, fun31.params, fun32.params, fun4.params, num.sim = 10,
  num.per = 8, num.day = 5, ...)
}
\arguments{
\item{fun2.params}{parameters for function 2}

\item{fun31.params}{parameters for first call of function 3}

\item{fun32.params}{parameters for second call of funtion 3}

\item{fun4.params}{parameters for call to function 4}

\item{num.sim}{first variable for function 1}

\item{num.per}{second variable for function 1}

\item{num.day}{third variable for function 1}
}
\description{
Test function level 1
}
