% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamfuncs.R
\name{fun2}
\alias{fun2}
\title{Test function level 2}
\usage{
fun2(var21 = 0.5, var22 = 5, ...)
}
\arguments{
\item{var21}{first variable for function 2}

\item{var22}{second variable for function 2}
}
\description{
Test function level 2
}
