% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamfuncs.R
\name{fun3}
\alias{fun3}
\title{Test function level 3}
\usage{
fun3(var31 = 1.396, var32 = 14.387, var33 = 3.219, fun3on = TRUE, ...)
}
\arguments{
\item{var31}{first variable for function 3}

\item{var32}{second variable for function 3}

\item{var33}{third variable for function 3}

\item{fun3on}{turn the formula on or off}
}
\description{
Test function level 3
}
