% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamfuncs.R
\name{localregargs}
\alias{localregargs}
\title{local regular arguments localregargs function}
\usage{
localregargs(var21, var22, fun22on, var31, fun3on, var41, fun4on)
}
\description{
local regular arguments localregargs function
}
